API/requirements.txt
aiohttp==3.10.5
APScheduler==3.11.0
Flask==3.1.0
IP2Location==8.10.5
mysql_connector_repackaged==0.3.1
Requests==2.32.3
SocksiPy_branch==1.01
tqdm==4.66.5
